refactor(parser): move scanner diagnostic predicates
